apps/user-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.1.2
prometheus-flask-exporter==0.23.0
gunicorn==21.2.0
